Move report printing to ReportingContext
