+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/instance/ReportingContext.java
@@ -19,19 +19,97 @@ package sleeper.systemtest.drivers.instance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import sleeper.systemtest.drivers.util.TestContext;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.io.PrintStream;
+import java.io.UncheckedIOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.time.Instant;
 
 public class ReportingContext {
     private static final Logger LOGGER = LoggerFactory.getLogger(ReportingContext.class);
 
+    private final SystemTestParameters parameters;
     private Instant recordingStartTime = Instant.now();
+
+    public ReportingContext(SystemTestParameters parameters) {
+        this.parameters = parameters;
+    }
 
     public void startRecording() {
         recordingStartTime = Instant.now();
         LOGGER.info("Ingest recording window started at {}", recordingStartTime);
     }
 
-    public Instant getRecordingStartTime() {
-        return recordingStartTime;
+    public void print(TestContext testContext, SystemTestReport report) {
+        try (ReportHandle handle = openReport(testContext)) {
+            report.print(handle.getPrintStream(), recordingStartTime);
+        }
+    }
+
+    private ReportHandle openReport(TestContext testContext) {
+        Path outputDirectory = parameters.getOutputDirectory();
+        if (outputDirectory != null) {
+            return new FileWriter(outputDirectory.resolve(
+                    testContext.getTestClassAndMethod() + ".report.log"));
+        } else {
+            return new LogWriter();
+        }
+    }
+
+    interface ReportHandle extends AutoCloseable {
+        PrintStream getPrintStream();
+
+        void close();
+    }
+
+    private static class FileWriter implements ReportHandle {
+        private final OutputStream outputStream;
+        private final PrintStream printStream;
+
+        FileWriter(Path file) {
+            try {
+                outputStream = Files.newOutputStream(file);
+                printStream = new PrintStream(outputStream, false, StandardCharsets.UTF_8);
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
+            }
+        }
+
+        @Override
+        public PrintStream getPrintStream() {
+            return printStream;
+        }
+
+        @Override
+        public void close() {
+            try {
+                printStream.flush();
+                outputStream.close();
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
+            }
+        }
+    }
+
+    private static class LogWriter implements ReportHandle {
+        private final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        private final PrintStream printStream = new PrintStream(outputStream, false, StandardCharsets.UTF_8);
+
+        @Override
+        public PrintStream getPrintStream() {
+            return printStream;
+        }
+
+        @Override
+        public void close() {
+            printStream.close();
+            LOGGER.info("Reports:\n{}", outputStream.toString(StandardCharsets.UTF_8));
+        }
     }
 }
